fix(bake): Fix error when updating artifacts after bake (#697)
* test(bake): Add test exercising bake artifact handling

There is currently no test that exercises the logic that updates the
name and uuid on artifacts produced by a bake.  Add a test exercising
this logic that currently fails and will be fixed in the next commit.

* fix(bake): Fix error when updating artifacts after bake

When artifacts were made immutable, this usage was missed as it
did not have tests (and was in groovy so compilation didn't catch
it). Instead of mutating the artifacts, return copies with the
appropriate fields changed.